+++ src/edges_analysis/cli.py
@@ -6,28 +6,79 @@ from pathlib import Path
 import yaml
 import h5py
 import tqdm
-from colored import stylize, fg, bg, attr
+from rich.console import Console
+from rich.logging import RichHandler
 from .analysis.levels import Level1
 from .analysis import levels
 from .config import config
-from edges_io.logging import logger
 import logging
 import numpy as np
 
+from rich.panel import Panel
+from rich import box
+from rich.rule import Rule
+from rich.table import Table
+
+console = Console()
+
 main = click.Group()
 
-logger.setLevel(logging.INFO)
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(message)s",
+    datefmt="[%X]",
+    handlers=[
+        RichHandler(
+            # console=console,
+            show_time=False,
+            show_path=False,
+            markup=True,
+            rich_tracebacks=True,
+            tracebacks_show_locals=True,
+        )
+    ],
+)
+
+logger = logging.getLogger(__name__)
 
 
-def _get_settings(settings):
+def _get_settings(settings, xrfi, **cli_settings):
     with open(settings, "r") as fl:
         settings = yaml.load(fl, Loader=yaml.FullLoader)
 
-    print(stylize("Settings:", attr("bold")))
+    settings.update(cli_settings)
+    if not xrfi and settings["xrfi_pipe"]:
+        settings["xrfi_pipe"] = {}
+
+    console.print("[bold]Settings:")
     for k, v in settings.items():
-        print(f"    {k}:", stylize(v, attr("dim")))
+        console.print(f"    {k}: [dim]{v}[/]")
 
     return settings
+
+
+def _ctx_to_dct(args):
+    dct = {}
+    j = 0
+    while j < len(args):
+        arg = args[j]
+        if "=" in arg:
+            a = arg.split("=")
+            dct[a[0].replace("--", "")] = a[-1]
+            j += 1
+        else:
+            dct[arg.replace("--", "")] = args[j + 1]
+            j += 2
+
+    for k in dct:
+        for tp in (int, float):
+            try:
+                dct[k] = tp(dct[k])
+                break
+            except TypeError:
+                pass
+
+    return dct
 
 
 def _get_input_files(level, path, label, allow_many=False):
@@ -40,10 +91,7 @@ def _get_input_files(level, path, label, allow_many=False):
         with open(label, "r") as fl:
             settings = yaml.load(fl, Loader=yaml.FullLoader)
             label = settings.pop("label", hashlib.md5(repr(settings).encode()).hexdigest())
-        print(
-            stylize("Using auto-generated label: ", attr("bold")),
-            stylize(label, attr("dim")),
-        )
+        console.print(f"[bold]Using auto-generated label: [dim]{label}[/]")
 
     fnc = _get_files if allow_many else _get_unique_file
 
@@ -63,10 +111,10 @@ def _get_input_files(level, path, label, allow_many=False):
         except ValueError:
             pass
 
-    print(stylize("Tried Following Paths: ", attr("bold")))
+    console.print("[bold]Tried Following Paths: ")
 
     for pth in try_paths:
-        print(stylize(f"\t{pth}", attr("dim")))
+        console.print(f"[dim]\t{pth}")
 
     return []
 
@@ -117,20 +165,34 @@ def get_output_path(level, settings, in_file, label, prefix):
         return prefix / in_file.name
 
 
-@main.command()
+@main.command(
+    context_settings={  # Doing this allows arbitrary options to override config
+        "ignore_unknown_options": True,
+        "allow_extra_args": True,
+    }
+)
 @click.argument("settings", type=click.Path(dir_okay=False, exists=True))
 @click.argument("path", type=click.Path(dir_okay=True), nargs=-1)
 @click.option("-l", "--label", default="")
 @click.option("-p", "--prefix", default="")
 @click.option("-m", "--message", default="")
+@click.option("-x/-X", "--xrfi/--no-xrfi", default=True, help="manually turn off xRFI")
 @click.option(
     "-c/-C",
     "--clobber/--no-clobber",
     help="whether to overwrite any existing data at the output location",
 )
-def calibrate(settings, path, label, prefix, message, clobber):
+def calibrate(ctx, settings, path, label, prefix, message, xrfi, clobber):
     """Calibrate field data to produce Level1 files."""
-    settings = _get_settings(settings)
+    console.print(
+        Panel("edges-analysis [blue]calibrate[/]", box=box.DOUBLE_EDGE),
+        style="bold",
+        justify="center",
+    )
+
+    cli_settings = _ctx_to_dct(ctx.args)
+    settings = _get_settings(settings, xrfi, **cli_settings)
+
     label = settings.pop("label", "") or label
 
     path = [Path(p) for p in path]
@@ -153,9 +215,8 @@ def calibrate(settings, path, label, prefix, message, clobber):
 
     output_dir = get_output_dir(1, settings, label, prefix)
 
-    print(
-        stylize("Output Directory:", attr("bold")),
-        stylize(str(output_dir), attr("dim")),
+    console.print(
+        f"[bold]Output Directory: [dim]{output_dir}",
     )
 
     # If the directory is not empty, we need to check whether the files that are
@@ -209,13 +270,17 @@ Metadata in file:
 
         l1.write(fname)
 
-    print(
-        stylize("All files written to:", attr("bold")),
-        stylize(str(output_dir), attr("dim")),
+    console.print(
+        f"[bold]All files written to: [dim]{output_dir}",
     )
 
 
-@main.command()
+@main.command(
+    context_settings={  # Doing this allows arbitrary options to override config
+        "ignore_unknown_options": True,
+        "allow_extra_args": True,
+    }
+)
 @click.argument("level", type=int)
 @click.argument("settings", type=click.Path(exists=True, dir_okay=False))
 @click.option("-i", "--path", default="", help="path to input files")
@@ -238,40 +303,53 @@ Metadata in file:
     default="",
     help="optional message to insert into the HDF5 file describing the reason for this analysis",
 )
+@click.option("-x/-X", "--xrfi/--no-xrfi", default=True, help="manually turn off xRFI")
 @click.option(
     "-c/-C",
     "--clobber/--no-clobber",
     help="whether to overwrite any existing data at the output location",
 )
-def level(level, settings, path, label, prev_label, prefix, message, clobber):
+@click.pass_context
+def level(ctx, level, settings, path, label, prev_label, prefix, message, xrfi, clobber):
     """Bump from a level to the next level."""
     assert level > 1
 
-    settings = _get_settings(settings)
+    console.print(
+        Panel(f"edges-analysis [blue]Level {level}[/]", box=box.DOUBLE_EDGE),
+        style="bold",
+        justify="center",
+    )
+
+    cli_settings = _ctx_to_dct(ctx.args)
+    settings = _get_settings(settings, xrfi, **cli_settings)
     label = settings.pop("label", "") or label
 
     in_files = _get_input_files(level - 1, path, prev_label, level in [2])
     if not in_files:
-        print(stylize("Found no input files!", fg("red") + attr("bold")))
+        console.print("[bold red]Found no input files!")
         return
 
     if isinstance(in_files, list):
-        print(stylize(f"Combining {len(in_files)} Level{level-1} files:", attr("bold")))
+        console.print(f"[bold]Combining {len(in_files)} Level{level-1} files:")
         for fl in in_files:
-            print(stylize(f"\t{fl.absolute()}", attr("dim")))
+            console.print(f"[blue]\t{fl.absolute()}")
     else:
-        print(stylize(f"Processing {in_files.name}", attr("bold")))
+        console.print(f"[bold]Processing[/] '{in_files.name}'")
 
     # Get the output structure ready
     output_file = get_output_path(level, settings, in_files, label, prefix)
 
-    print(stylize("Output File:", attr("bold")), stylize(str(output_file), attr("dim")))
-
     if output_file.exists() and not clobber:
-        logger.error("The output file already exists -- use clobber!")
+        logger.error(
+            f"[bold red]The output file [blue]'{output_file}'[/] already exists -- use clobber!"
+        )
         return
 
+    console.print()
+    console.print(Rule("Beginning Level Upgrade"))
     getattr(levels, f"Level{level}").from_previous_level(in_files, output_file, clobber, **settings)
 
     with h5py.File(output_file, "a") as fl:
         fl.attrs["message"] = message
+
+    console.print(f"[bold]Output File: [blue]{output_file}")